Ensure all updates to user preference happens from the cache
