+++ src/dotnet/APIView/APIViewWeb/Managers/UserProfileManager.cs
@@ -5,7 +5,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Security.Claims;
 using System.Threading.Tasks;
-using APIView.Identity;
 using APIViewWeb.Models;
 using APIViewWeb.Repositories;
 
@@ -33,22 +32,6 @@ namespace APIViewWeb.Managers
         public async Task UpdateUserProfile(string userName, UserProfileModel profile)
         {
             await _UserProfileRepository.UpsertUserProfileAsync(userName, profile);
-        }
-
-        public async Task SetUserEmailIfNullOrEmpty(ClaimsPrincipal User)
-        {
-            var userProfile = await TryGetUserProfileAsync(User);
-
-            if (string.IsNullOrWhiteSpace(userProfile.Email))
-            {
-                var microsoftEmail = User.Claims.FirstOrDefault(c => c.Type == ClaimConstants.Email)?.Value;
-
-                if (!string.IsNullOrWhiteSpace(microsoftEmail))
-                {
-                    userProfile.Email = microsoftEmail;
-                    await UpdateUserProfile(User.GetGitHubLogin(), userProfile);
-                }
-            }
         }
     }
 }
